requirements.txt
streamlit>=1.32
plotly>=5.18
pandas>=2.0
pyarrow>=14.0
python-dotenv>=1.0
db-dtypes>=1.2
google-cloud-bigquery>=3.10
google-cloud-bigquery-storage>=2.20

